Rename exchangedTokenRehydrator and type safe function
